requirements.txt
opencv-python
matplotlib
tqdm
jupyter
git+https://github.com/openai/CLIP.git
pytorch-lightning
scipy
lpips
torch-fidelity